+++ octopod/ensemble/dataset.py
@@ -1,5 +1,6 @@
 import numpy as np
 from PIL import Image
+from sklearn import preprocessing
 import torch
 from torch.utils.data import Dataset
 
@@ -18,7 +19,8 @@ class OctopodEnsembleDataset(Dataset):
     img_inputs: pandas Series
         the paths to images to be used
     y: list
-        A list of lists of dummy-encoded categories
+        A list of dummy-encoded categories or strings,
+        which will be encoded using a sklearn label encoder
     tokenizer: pretrained BERT Tokenizer
         BERT tokenizer likely from `transformers`
     max_seq_length: int (defaults to 128)
@@ -45,6 +47,7 @@ class OctopodEnsembleDataset(Dataset):
         self.y = y
         self.tokenizer = tokenizer
         self.max_seq_length = max_seq_length
+        self.label_encoder, self.label_mapping = self._encode_labels()
 
         if transform == 'train' or 'val':
             self.transform = full_img_transforms[transform]
@@ -83,7 +86,10 @@ class OctopodEnsembleDataset(Dataset):
         full_img = self.transform(full_img)
         cropped_img = self.crop_transform(cropped_img)
 
-        y_output = torch.from_numpy(np.array(self.y[index])).long()
+        label = self.y[index]
+        label = self.label_encoder.transform([label])[0]
+
+        y_output = torch.from_numpy(np.array(label)).long()
 
         return {'bert_text': bert_text,
                 'full_img': full_img,
@@ -91,6 +97,13 @@ class OctopodEnsembleDataset(Dataset):
 
     def __len__(self):
         return len(self.text_inputs)
+
+    def _encode_labels(self):
+        """Encodes string or numeric y labels to integers using LabelEncoder"""
+        le = preprocessing.LabelEncoder()
+        le.fit(self.y)
+        mapping_dict = dict(zip(le.transform(le.classes_), le.classes_))
+        return le, mapping_dict
 
 
 class OctopodEnsembleDatasetMultiLabel(OctopodEnsembleDataset):
@@ -104,9 +117,11 @@ class OctopodEnsembleDatasetMultiLabel(OctopodEnsembleDataset):
     img_inputs: pandas Series
         the paths to images to be used
     y: list
-        a list of binary encoded categories with length equal to number of
+        a list of lists of binary encoded categories or strings with length equal to number of
         classes in the multi-label task. For a 4 class multi-label task
-        a sample list would be [1,0,0,1]
+        a sample list would be [1,0,0,1], A string example would be ['cat','dog'],
+        (if the classes were ['cat','frog','rabbit','dog]), which will be encoded
+        using a sklearn label encoder to [1,0,0,1].
     tokenizer: pretrained BERT Tokenizer
         BERT tokenizer likely from `transformers`
     max_seq_length: int (defaults to 128)
@@ -147,8 +162,19 @@ class OctopodEnsembleDatasetMultiLabel(OctopodEnsembleDataset):
         full_img = self.transform(full_img)
         cropped_img = self.crop_transform(cropped_img)
 
-        y_output = torch.FloatTensor(self.y[index])
+        label = self.y[index]
+        label = list(self.label_encoder.transform([label])[0])
+
+        y_output = torch.FloatTensor(label)
 
         return {'bert_text': bert_text,
                 'full_img': full_img,
                 'crop_img': cropped_img}, y_output
+
+    def _encode_labels(self):
+        """Encodes string or numeric y labels to integers using MultiLabelBinarizer"""
+        mlb = preprocessing.MultiLabelBinarizer()
+        mlb.fit(self.y)
+        mapping_dict = dict(zip(list(range(0, len(mlb.classes_))), mlb.classes_))
+
+        return mlb, mapping_dict